requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LLM & Agent Framework
langchain==0.1.4
langchain-community==0.0.10
langgraph==0.0.15
langsmith==0.1.0

# Vector Database & Embeddings
chromadb==0.4.19
sentence-transformers==2.2.2
openai==1.3.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
PyPDF2==4.0.1
python-docx==0.8.11

# Evaluation & Metrics
ragas==0.1.2
scikit-learn==1.3.2
nltk==3.8.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
aiohttp==3.9.1
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# AWS (Optional for cloud deployment)
boto3==1.29.7
botocore==1.32.7
